backend/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cmake==3.30.3
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
flatbuffers==24.3.25
fonttools==4.53.1
gunicorn==23.0.0
idna==3.10
jax==0.4.33
jaxlib==0.4.33
kiwisolver==1.4.7
matplotlib==3.9.2
mediapipe==0.10.14
ml_dtypes==0.5.0
numpy==2.1.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
paypalrestsdk==1.13.3
pillow==10.4.0
prettytable==3.11.0
protobuf==4.25.5
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.14.1
six==1.16.0
sounddevice==0.5.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
wcwidth==0.2.13
whitenoise==6.8.2
